Move Func GPS to UserSettingFragment
