+++ app/src/main/java/kr/co/ajjulcoding/team/project/holo/HomeViewModel.kt
@@ -10,23 +10,9 @@ import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
 
 class HomeViewModel: ViewModel() {
-    private var _userLocation = MutableLiveData<String>()
-    val userLocation: LiveData<String> = _userLocation
-    private var _userProile = MutableLiveData<Uri>()
-    val userProfile:LiveData<Uri> = _userProile
     private var _chatRoom = MutableLiveData<ChatRoom>()
     val chatRoom:LiveData<ChatRoom> = _chatRoom
-//    private var _validChat = MutableLiveData<Boolean>()
-//    val validChat: LiveData<Boolean> = _validChat
     val repository = Repository()
-
-    fun setUserLocation(location:String){
-        _userLocation.value = location
-    }
-
-    fun setUserProfile(imgUri: Uri){
-        _userProile.value = imgUri
-    }
 
     fun getUserNicknameAndToken(email:String)= viewModelScope.async {
         return@async repository.getUserNicknameAndToken(email)
